server/requirements.txt
fastapi
uvicorn
python-multipart
minio
psycopg2-binary
httpx
websockets
numpy
scipy
matplotlib
opencv-python-headless
torch
albumentations
timm
scikit-learn
scikit-image
pandas
pywavelets
python-jose[cryptography]